Move propertypages into diagram package
That's the last bit. Now the adapters package is gone.
