--- gaphor/adapters/__init__.py
@@ -1,20 +1,0 @@
-import gaphor.diagram.editors
-import gaphor.diagram.grouping
-
-import gaphor.diagram.components.connectorconnect
-import gaphor.diagram.interactions.messageconnect
-import gaphor.diagram.profiles.extensionconnect
-import gaphor.diagram.profiles.stereotypepage
-import gaphor.diagram.profiles.metaclasseditor
-
-import gaphor.diagram.states.vertexconnect
-import gaphor.diagram.states.propertypages
-
-
-import gaphor.diagram.actions.flowconnect
-import gaphor.diagram.actions.partitionpage
-
-import gaphor.diagram.classes.classconnect
-import gaphor.diagram.classes.interfaceconnect
-
-import gaphor.diagram.usecases.usecaseconnect
