+++ gaphor/diagram/actions/__init__.py
@@ -1,12 +1,19 @@
-from .action import ActionItem
+from .action import ActionItem, SendSignalActionItem, AcceptEventActionItem
 from .activitynodes import (
     ActivityNodeItem,
     ActivityFinalNodeItem,
     FlowFinalNodeItem,
+    ForkNodeItem,
     DecisionNodeItem,
+    InitialNodeItem,
 )
 from .flow import FlowItem
 from .objectnode import ObjectNodeItem
 from .partition import PartitionItem
 
-# from . import actionsgrouping, flowconnect, partitionpage
+
+def _load():
+    from . import actionsgrouping, flowconnect, partitionpage, actionspropertypages
+
+
+_load()
