+++ gaphor/diagram/classes/__init__.py
@@ -6,4 +6,9 @@ from .interface import InterfaceItem
 from .klass import ClassItem
 from .package import PackageItem
 
-# from . import classconnect, interfaceconnect
+
+def _load():
+    from . import classconnect, interfaceconnect, classespropertypages
+
+
+_load()
