+++ gaphor/diagram/general/__init__.py
@@ -2,4 +2,9 @@ from .comment import CommentItem
 from .commentline import CommentLineItem
 from .simpleitem import Box, Ellipse, Line
 
-from . import connectors
+
+def _load():
+    from . import connectors, generalpropertypages
+
+
+_load()
