+++ gaphor/diagram/interactions/__init__.py
@@ -2,4 +2,9 @@ from .interaction import InteractionItem
 from .message import MessageItem
 from .lifeline import LifelineItem
 
-# from . import messageconnect, interactionsgrouping
+
+def _load():
+    from . import messageconnect, interactionsgrouping, interactionspropertypages
+
+
+_load()
