+++ gaphor/diagram/items.py
@@ -11,56 +11,11 @@ from .nameditem import NamedItem
 from .compartment import CompartmentItem, FeatureItem
 from .classifier import ClassifierItem
 
-# General:
-from .general.comment import CommentItem
-from .general.commentline import CommentLineItem
-from .general.simpleitem import Line, Box, Ellipse
-
-# Classes:
-from .classes.klass import ClassItem, OperationItem
-from .classes.interface import InterfaceItem
-from .classes.package import PackageItem
-from .classes.association import AssociationItem
-from .classes.dependency import DependencyItem
-from .classes.generalization import GeneralizationItem
-from .classes.implementation import ImplementationItem
-
-# Components:
-from .components.artifact import ArtifactItem
-from .components.connector import ConnectorItem
-from .components.component import ComponentItem
-from .components.node import NodeItem
-from .components.subsystem import SubsystemItem
-
-# Actions:
-from .actions.activitynodes import ActivityNodeItem
-from .actions.activitynodes import InitialNodeItem, ActivityFinalNodeItem
-from .actions.activitynodes import FlowFinalNodeItem
-from .actions.activitynodes import DecisionNodeItem
-from .actions.activitynodes import ForkNodeItem
-from .actions.objectnode import ObjectNodeItem
-from .actions.action import ActionItem, SendSignalActionItem, AcceptEventActionItem
-from .actions.flow import FlowItem
-from .actions.partition import PartitionItem
-
-# Interactions
-from .interactions.interaction import InteractionItem
-from .interactions.lifeline import LifelineItem
-from .interactions.message import MessageItem
-
-# States
-from .states.state import VertexItem
-from .states.state import StateItem
-from .states.transition import TransitionItem
-from .states.finalstate import FinalStateItem
-from .states.pseudostates import InitialPseudostateItem, HistoryPseudostateItem
-
-# Use Cases:
-from .usecases.actor import ActorItem
-from .usecases.usecase import UseCaseItem
-from .usecases.include import IncludeItem
-from .usecases.extend import ExtendItem
-
-# Stereotypes:
-from .profiles.extension import ExtensionItem
-from .profiles.metaclass import MetaclassItem
+from .actions import *
+from .classes import *
+from .components import *
+from .general import *
+from .interactions import *
+from .profiles import *
+from .states import *
+from .usecases import *
