+++ gaphor/diagram/usecases/__init__.py
@@ -3,4 +3,9 @@ from .extend import ExtendItem
 from .include import IncludeItem
 from .usecase import UseCaseItem
 
-from . import usecaseconnect
+
+def _load():
+    from . import usecaseconnect
+
+
+_load()
