+++ gaphor/ui/mainwindow.py
@@ -34,7 +34,6 @@ from gaphor.ui.layout import deserialize
 from gaphor.ui.namespace import Namespace
 from gaphor.ui.toolbox import Toolbox
 
-import gaphor.adapters
 
 log = logging.getLogger(__name__)
